split module `print` into sub-modules for docs and code, move remaining print-related functions from document to print
